FactionMember: move code around
