+++ client/windows/CHeroWindow.h
@@ -9,6 +9,8 @@
  */
 #pragma once
 
+#include <vcmi/FactionMember.h>
+
 #include "../../lib/HeroBonus.h"
 #include "../widgets/CWindowWithArtifacts.h"
 #include "../widgets/CGarrisonInt.h"
@@ -45,7 +47,7 @@ public:
 };
 
 //helper class for calculating values of hero bonuses without bonuses from picked up artifact
-class CHeroWithMaybePickedArtifact : public virtual IBonusBearer
+class CHeroWithMaybePickedArtifact : public IBonusBearer, public IFactionMember
 {
 public:
 	const CGHeroInstance * hero;
@@ -53,6 +55,9 @@ public:
 
 	CHeroWithMaybePickedArtifact(CWindowWithArtifacts * Cww, const CGHeroInstance * Hero);
 	TConstBonusListPtr getAllBonuses(const CSelector & selector, const CSelector & limit, const CBonusSystemNode * root = nullptr, const std::string & cachingStr = "") const override;
+
+	const IBonusBearer * getBonusBearer() const override;
+	FactionID getFaction() const override; 
 
 	int64_t getTreeVersion() const override;
