+++ include/vcmi/Entity.h
@@ -31,43 +31,6 @@ public:
 	virtual bool isNativeTerrain(Identifier<ETerrainId> terrain) const;
 };
 
-class DLL_LINKAGE IFactionMember: public IConstBonusProvider, public INativeTerrainProvider
-{
-public:
-	/**
-	 Returns native terrain considering some terrain bonuses.
-	*/
-	virtual Identifier<ETerrainId> getNativeTerrain() const;
-	/**
-	 Returns magic resistance considering some bonuses.
-	*/
-	virtual int32_t magicResistance() const;
-	/**
-	 Returns minimal damage of creature or (when implemented) hero.
-	*/
-	virtual int getMinDamage(bool ranged) const;
-	/**
-	 Returns maximal damage of creature or (when implemented) hero.
-	*/
-	virtual int getMaxDamage(bool ranged) const;
-	/**
-	 Returns attack of creature or hero.
-	*/
-	virtual int getAttack(bool ranged) const;
-	/**
-	 Returns defence of creature or hero.
-	*/
-	virtual int getDefense(bool ranged) const;
-};
-
-/// Base class for creatures and battle stacks
-class DLL_LINKAGE ICreature: public IFactionMember
-{
-public:
-	ui32 Speed(int turn = 0, bool useBind = false) const; //get speed (in moving tiles) of creature with all modificators
-	ui32 MaxHealth() const; //get max HP of stack with all modifiers
-};
-
 class DLL_LINKAGE Entity
 {
 public:
@@ -93,11 +56,6 @@ public:
 
 template <typename IdType>
 class DLL_LINKAGE EntityWithBonuses : public EntityT<IdType>, public IConstBonusProvider
-{
-};
-
-template <typename IdType>
-class DLL_LINKAGE CreatureEntity : public EntityT<IdType>, public ICreature
 {
 };
